Computing degree and distance based topological indices of graphs
Computing various distance-based topological indices of graphs using newGraph and Matlab.

Let G be a graph for which you would like to compute a degree and distance based topological index.

Step 1. Draw G on newGraph and check distance matrix as an invariant by going to "Invariants" tab to "Select Invariant".

Copy the distance matrix by right clicking on the matrix.

Step 2. Paste the matrix and order of G in the indices.m file in B and n respectively.

Run the indices.m file to obtain your result.